git69 split connectors
